// src/types/ticket-detail.ts
export type TicketDetail = {
  ticketCode: string;
  isPaid: boolean;
  userName: string;
  userPhone: string;
  destinationName: string;
  pickupLocationName: string;
  destinationPrice: number;
  quantity: number;
  totalPrice: number;
  date: string;
  departureTime: string;
  returnTime: string;
  createdAt: string;
};